added krishnan's code from latest PR
