+++ my-app/src/App.tsx
@@ -4,13 +4,14 @@ import "./App.css";
 import ForgotPasswordPage from "./pages/Login/forgot-passsowrd"; // Assuming this is the component for forgot password
 import Login from "./pages/Login/Login";
 import Dashboard from "./pages/Dashboard";
-
+import CalendarPage from "./components/Calendar/CalendarPage";
 function App() {
   return (
     <Router>
       <Routes>
         <Route path="/" element={<Login />} />
         <Route path="/forgot-password" element={<ForgotPasswordPage />} />
+        <Route path="/calendar" element={<CalendarPage />} />
         <Route path="/dashboard" element={<Dashboard />} />
       </Routes>
     </Router>
